Move duplicate tests to the parent abstract class
